consumer/requirements.txt
psycopg2-binary
datetime 
kafka-python
python-json-logger